requirements.txt
streamlit>=1.38.0
google-generativeai>=0.8.3
plotly>=5.24.1
pandas>=2.2.3
pillow>=10.4.0
streamlit-extras>=0.4.0
python-dotenv>=1.0.1
openai>=1.12.0
anthropic>=0.18.0
cohere>=4.0.0
huggingface-hub>=0.20.0
